fix(fe:FSADT1-539): - refactor form section schemas, form initial data, functions into different folders refs: #89
